requirements.txt
asgiref
attrs
autobahn
Automat
certifi
cffi
channels
charset-normalizer
constantly
cryptography
daphne
Django
django-admin-honeypot-updated-2021
django-cors-headers
django-environ
django-filter
django-ipware
django-unfold
django-waitress
djangorestframework
hyperlink
idna
incremental
Markdown
pyasn1
pyasn1_modules
pycparser
pyOpenSSL
python-ipware
requests
service-identity
sqlparse
Twisted
txaio
typing_extensions
tzdata
urllib3
waitress
zope.interface